SOLR-17518: Deprecate UpdateRequest.getXml() and replace it with XMLRequestWriter (#3200)
