+++ solr/core/src/test/org/apache/solr/update/AddBlockUpdateTest.java
@@ -54,6 +54,7 @@ import org.apache.lucene.search.join.QueryBitSetProducer;
 import org.apache.lucene.search.join.ScoreMode;
 import org.apache.lucene.search.join.ToParentBlockJoinQuery;
 import org.apache.solr.SolrTestCaseJ4;
+import org.apache.solr.client.solrj.impl.XMLRequestWriter;
 import org.apache.solr.client.solrj.request.RequestWriter;
 import org.apache.solr.client.solrj.request.UpdateRequest;
 import org.apache.solr.common.SolrException;
@@ -451,7 +452,7 @@ public class AddBlockUpdateTest extends SolrTestCaseJ4 {
     Collections.shuffle(docs, random());
     req.add(docs);
 
-    RequestWriter requestWriter = new RequestWriter();
+    RequestWriter requestWriter = new XMLRequestWriter();
     OutputStream os = new ByteArrayOutputStream();
     requestWriter.write(req, os);
     assertBlockU(os.toString());
@@ -518,7 +519,7 @@ public class AddBlockUpdateTest extends SolrTestCaseJ4 {
     Collections.shuffle(docs, random());
     req.add(docs);
 
-    RequestWriter requestWriter = new RequestWriter();
+    RequestWriter requestWriter = new XMLRequestWriter();
     OutputStream os = new ByteArrayOutputStream();
     requestWriter.write(req, os);
     assertBlockU(os.toString());
@@ -704,7 +705,7 @@ public class AddBlockUpdateTest extends SolrTestCaseJ4 {
     Collections.shuffle(docs, random());
     req.add(docs);
 
-    RequestWriter requestWriter = new RequestWriter();
+    RequestWriter requestWriter = new XMLRequestWriter();
     OutputStream os = new ByteArrayOutputStream();
     requestWriter.write(req, os);
     assertBlockU(os.toString());
@@ -807,7 +808,7 @@ public class AddBlockUpdateTest extends SolrTestCaseJ4 {
     Collections.shuffle(docs, random());
     req.add(docs);
 
-    RequestWriter requestWriter = new RequestWriter();
+    RequestWriter requestWriter = new XMLRequestWriter();
     OutputStream os = new ByteArrayOutputStream();
     requestWriter.write(req, os);
     assertBlockU(os.toString());
