+++ solr/solrj/src/java/org/apache/solr/client/solrj/util/ClientUtils.java
@@ -17,7 +17,6 @@
 package org.apache.solr.client.solrj.util;
 
 import java.io.IOException;
-import java.io.StringWriter;
 import java.io.Writer;
 import java.net.MalformedURLException;
 import java.net.URI;
@@ -210,15 +209,6 @@ public class ClientUtils {
         XML.writeXML(writer, "field", valWriter, "name", name, "update", update);
       }
     }
-  }
-
-  public static String toXML(SolrInputDocument doc) {
-    StringWriter str = new StringWriter();
-    try {
-      writeXML(doc, str);
-    } catch (Exception ex) {
-    }
-    return str.toString();
   }
 
   // ---------------------------------------------------------------------------------------
